saving loaded data to db. get it while offline

+++ app/src/main/java/com/kerer/weatherapp/api/dto/CurrentlyDTO.java
@@ -47,7 +47,7 @@ public class CurrentlyDTO {
     private Double visibility;
     @SerializedName("cloudCover")
     @Expose
-    private Integer cloudCover;
+    private Double cloudCover;
     @SerializedName("pressure")
     @Expose
     private Double pressure;
@@ -159,11 +159,11 @@ public class CurrentlyDTO {
         this.visibility = visibility;
     }
 
-    public Integer getCloudCover() {
+    public Double getCloudCover() {
         return cloudCover;
     }
 
-    public void setCloudCover(Integer cloudCover) {
+    public void setCloudCover(Double cloudCover) {
         this.cloudCover = cloudCover;
     }
